% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shr_list_variables.R
\name{odr_list_variables}
\alias{odr_list_variables}
\alias{shr_list_variables}
\title{Get informations related to the variables available for a given collection}
\usage{
shr_list_variables(collection)
}
\value{
A data.frame with the available variables for the collection, and a set of related information for each variable.
}
\description{
Get the variables available for a given collection
}
\examples{

# Get the variables available for the collection S2L2A
(df_varinfo <- shr_list_variables("S2L2A"))


}
