% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shr_get_url.R
\name{shr_get_url}
\alias{shr_get_url}
\title{Get URLs of Sentinel datasets}
\usage{
shr_get_url(collection, variables, roi, time_range,
  cloud_cover_max = NULL, instance_id = NULL, verbose = FALSE)
}
\value{
a data.frame with one row for each dataset and 3 columns  :
 \itemize{
 \item{*time_start*: }{Start Date/time for the dataset}
 \item{*name*: }{An indicative name for the dataset}
 \item{*url*: }{WMS URL for the dataset}
 }
}
\description{
This function enables to retrieve WMS URLs of Sentinel products given a collection, a ROI, a time frame and a set of bands of interest.
}
\details{
Available collections : S2L2A, S1-AWS-IW-VVVH

Available dimensions :
\itemize{
\item{for collection S2L2A  : } {1_TRUE_COLOR, 2_FALSE_COLOR, 3_NDVI, 4_FALSE_COLOR, 5_MOISTURE_INDEX, 6_SWIR, 7-NDWI, 8_NDSI, 9_SCENE_CLASSIFICATION, B01, B02, etc... B12}
\item{for collection S1-AWS-IW-VVVH : } {towrite}
}
}
\examples{

\dontrun{
# login to the wms server
log_s1 <- shr_login(Sys.getenv("instance_id_shub_s1"),"sentinel1")
log_s2 <- shr_login(Sys.getenv("instance_id_shub_s2"),"sentinel2")

roi <- sf::st_as_sf(data.frame(geom="POLYGON ((-5.82 9.54, -5.42 9.55, -5.41 8.84, -5.81 8.84, -5.82 9.54))"),wkt="geom",crs = 4326)
time_range<-as.Date(c("2017-01-01","2017-01-30"))

(s1_urls <-shr_get_url(collection = "S1-AWS-IW-VVVH", variables = c("VV","VH") ,roi = roi, time_range = time_range))

(s2l2a_urls <-shr_get_url(collection = "S2L2A", variables = c("B04","B08","B8A","B11","9_SCENE_CLASSIFICATION") ,roi = roi, time_range = time_range))
}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
