% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shr_list_collections.R
\name{shr_list_collections}
\alias{shr_list_collections}
\title{Get data sources / collections implemented in the package}
\usage{
shr_list_collections()
}
\value{
a data.frame with the data sources / collections dealt by shr_list_collections, along with details on each data collection
}
\description{
This function returns a table of the data sources / collections that can be downloaded using the shub4r package
}
\examples{

sources<-shr_list_collections()
sources

}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
