% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shr_login.R
\name{shr_login}
\alias{shr_login}
\title{Login to query servers and download data}
\usage{
shr_login(instance_id, source, verbose = TRUE)
}
\arguments{
\item{instance_id}{instance id}

\item{source}{source. See details}
}
\value{
None.
}
\description{
Login before querying data servers
}
\details{
Current options for parameter \code{"source"} are :
\itemize{
\item{"sentinel1"}{ : to query and download Sentinel 1 collections}
\item{"sentinel2"}{ : to query and download Sentinel 2 collections}
}
Create an account to the Sentinel Hub OGC web services here : \url{https://sentinel-hub.com/pricing-plans}.
}
\examples{

\donttest{
instance_id_s1 <- Sys.getenv("instance_id_shub_s1")
instance_id_s2 <- Sys.getenv("instance_id_shub_s2")

shr_login(instance_id_s1,source = "sentinel1")
shr_login(instance_id_s2,source = "sentinel2")
}

}
