% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shr_list_products.R
\name{shr_list_products}
\alias{shr_list_products}
\title{Get products available for a given collection, roi and time range}
\usage{
shr_list_products(collection, roi, time_range, instance_id = NULL)
}
\description{
Get products available for a given collection, roi and time range
}
